Move standalone openstack functions to an openstack module
These functions are easy to move and are specific to openstack.
This work is done in preparation of refactoring to support multiple
services.
